# src/static/api/api.yml
openapi: 3.0.0
info:
  title: Biology Database API
  description: 'Api rest to register users in a database for genetic biological applications. Go to [Home page](http://localhost:5000/) for more information.'
  version: 1.0.0
  termsOfService: 'http://localhost:5000'
  contact:
    email: rodrjuniorsantos@gmail.com
servers:
  - url: 'http://localhost:5000'
    description: 'Primary server, e.g. Main (production) server'
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: api-token
paths:
  /api/users/user:
    post:
      summary: Create a user Function. It will return the JWT token if the request was successful
      tags:
        - User
      requestBody:
        description: User Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - role
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
  
      responses:
        '200':
          description: User successfully registered
        '400':
          description: User already exist, please supply another email address
  /api/users/user_all:
    get:
      summary: Get all users Function
      security:
        - APIKeyHeader: []
      tags:
        - User
      
      responses:
        '200':
          description: Returns all users
        '400':
          description: User not found
        '401':
          description: Permission denied
  /api/users/{user_id}:
    get:
      summary: Gets a user by ID.
      security:
        - APIKeyHeader: []
      tags:
        - User
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The user ID.

      responses:
        '200':
          description: Data this user successfully
        '400':
          description: user not found
        '401':
          description: Permission denied
  /api/users/edit/{me}:
    put:
      summary: Update A User.
      security:
        - APIKeyHeader: []
      tags:
        - User
      parameters:
        - in: path
          name: me
          required: true
          schema:
            type: string
            description: The user ID.
      requestBody:
        description: User Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - role
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: User successfully update
        '400':
          description: Missing data
  /api/users/delete/me:
    delete:
      summary: Return your user data.
      security:
        - APIKeyHeader: []
      tags:
        - User
      responses:
        '200':
          description: An Access Token API to be used in Boleto Viewer
  /api/users/login:
    post:
      summary: 'User Login Function In API, return a token API.'
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: An Access Token API to be used in Boleto Viewer
        '400':
          description: you need email and password to sign in
        '401':
          description: you need email and password to sign in
        '402':
          description: Credentials are not valid
        '403':
          description: Credentials are not valid
  /api/users/me:
    get:
      summary: Return your user data.
      security:
        - APIKeyHeader: []
      tags:
        - User
      responses:
        '200':
          description: An Access Token API to be used in Boleto Viewer
        '400':
          description: user in post not found
...